PhotonNetworkingSyncing.
This is a basic photon implementation example where I have submitted the assets that will be demonstrating how the room is created and joined. And mostly it have 3 scripts that can be used most frequently for synchronizing the objects across the networks.
Three scripts are to sync the object position, then to sync the object animation, and another is to sync the vfx i.e particle effects.